+++ cmd/main.go
@@ -101,6 +101,7 @@ func Main[B firecore.Block](chain *firecore.Chain[B]) {
 	registerCommonFlags(chain)
 	apps.RegisterReaderNodeApp(chain, rootLog)
 	apps.RegisterReaderNodeStdinApp(chain, rootLog)
+	apps.RegisterReaderNodeFirehoseApp(chain, rootLog)
 	apps.RegisterMergerApp(rootLog)
 	apps.RegisterRelayerApp(rootLog)
 	apps.RegisterFirehoseApp(chain, rootLog)